requirements.txt
# Core Dependencies
requests>=2.31.0
tqdm>=4.66.0
urllib3>=2.2.0
pyyaml>=6.0.1

# Web Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
jinja2>=3.1.3
python-multipart>=0.0.6
httpx>=0.26.0           # Async HTTP client for health checks

# Security & Rate Limiting
slowapi>=0.1.9          # Rate limiting for FastAPI
tenacity>=8.2.0         # Retry logic with exponential backoff

# Optional LLM SDKs (install only if you use them)
google-generativeai>=0.3.0
ollama>=0.1.9
aiosqlite>=0.19.0
aiohttp>=3.9.0
openai>=1.6.0
anthropic>=0.18.0       # Anthropic Claude API
langgraph>=0.0.60       # LangGraph for agent workflows (optional)
langchain-core>=0.1.0   # LangChain core (required by langgraph)

# PDF Report Generation (optional)
xhtml2pdf>=0.2.13  # Recommended for Windows - pure Python
# pdfkit>=1.0.0  # Alternative - requires wkhtmltopdf installed

# Proxy Support
aiohttp-socks>=0.8.4  # SOCKS proxy support for aiohttp

# Enhanced DNS Resolution
dnspython>=2.4.0  # Full DNS record resolution
